% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{ad_archive}
\alias{ad_archive}
\title{ad_archive}
\usage{
ad_archive(file, fold_path = default_fold(file))
}
\arguments{
\item{file}{The file of interest}

\item{fold_path}{The folder location for the data archive}
}
\value{
nothing
}
\description{
Save a timestamped file of the MRV, useful for manual edits
}
