% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{ad_init}
\alias{ad_init}
\title{ad_init}
\usage{
ad_init(file, fold_path = default_fold(file))
}
\arguments{
\item{file}{The file of interest}

\item{fold_path}{The folder location for the archD}
}
\value{
Nothing
}
\description{
Initialize file: Use on an existing file to track and create to repository folder.
}
