% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{ad_lsv}
\alias{ad_lsv}
\title{ad_lsv}
\usage{
ad_lsv(file, fold_path = default_fold(file), look_back = 0)
}
\arguments{
\item{file}{file}

\item{fold_path}{fold}

\item{look_back}{positive integer. zero is the lsv, higher numbers for earlier versions}
}
\value{

}
\description{
returns version name
}
