% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{ad_save}
\alias{ad_save}
\title{ad_save}
\usage{
ad_save(x, file, fold_path = default_fold(file))
}
\arguments{
\item{file}{The file of interest}

\item{fold_path}{The folder location for the data archive}
}
\value{
nothing
}
\description{
Save an R object to file
}
