% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{ad_to_date}
\alias{ad_to_date}
\title{ad_to_date}
\usage{
ad_to_date(file, fold_path = default_fold(file))
}
\arguments{
\item{file}{The file of interest}

\item{fold_path}{The folder location for the data archive}
}
\value{
nothing
}
\description{
Check if archd is up to date
MRV: most recent version in the "working directory"
LSV: last saved version in the archd repo
}
