% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{add_flag}
\alias{add_flag}
\title{add_flag}
\usage{
add_flag(df, id, new_flag)
}
\arguments{
\item{df}{The dataframe}

\item{id}{id}

\item{new_flag}{}
}
\description{
Only to be used with Cmic dataset with a New_ID column
}
