% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{coldesc_dt}
\alias{coldesc_dt}
\title{Like coldesc for DT rendering in markdown}
\usage{
coldesc_dt(df)
}
\arguments{
\item{df}{data}
}
\value{
DT
}
\description{
Like coldesc for DT rendering in markdown
}
