% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_sim_df.R
\name{comp_btw}
\alias{comp_btw}
\title{comp_btw}
\usage{
comp_btw(df1, df2, id = NULL)
}
\arguments{
\item{df1}{data.frame}

\item{df2}{data.frame}

\item{id}{character}
}
\value{
differences
}
\description{
comp_btw
}
