% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_sim_df.R
\name{comp_sim_df}
\alias{comp_sim_df}
\title{comp_sim_df}
\usage{
comp_sim_df(df1, df2)
}
\arguments{
\item{df1}{data.frame}

\item{df2}{data.frame}
}
\value{
dataframe with details about differences
}
\description{
comp_sim_df
}
