% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_metadata.R
\name{create_project_file_tracking}
\alias{create_project_file_tracking}
\title{create_project_file_tracking}
\usage{
create_project_file_tracking(dir = NULL, add_profile = FALSE)
}
\arguments{
\item{dir}{chr In which folder should the file be created. Defaults to project folder.}

\item{add_profile}{logical Should a note be done in .Rprofile for non-default dir}
}
\description{
Create tracking file
}
