% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{default_fold}
\alias{default_fold}
\title{default_fold}
\usage{
default_fold(file)
}
\arguments{
\item{file}{The file of interest}
}
\value{
default archD repo path
}
\description{
suggested folder to create the archive. Default option is to have an "archD" repo in the project folder with a mirrored folder structure.
}
