% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{export_report}
\alias{export_report}
\title{Export report}
\usage{
export_report(plots, file, plots_per_page = 24)
}
\arguments{
\item{plots}{list of ggplots}

\item{file}{path}

\item{plots_per_page}{24}
}
\description{
Mostly for internal use. Does the page layout and exports figures.
}
