% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nearest.R
\name{extract_nearest}
\alias{extract_nearest}
\title{Extract from nearest available pixel}
\usage{
extract_nearest(x, y, max_range = NULL)
}
\arguments{
\item{x}{Raster* object}

\item{y}{points represented by a sf object}

\item{max_range}{numeric maximum distance in meters. Default is 5 x res}
}
\value{
A vector for RasterLayer objects, and a matrix for RasterStack or RasterBrick objects.
}
\description{
takes Raster* and assigns proper method
}
