% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nearest.R
\name{extract_nearest_value}
\alias{extract_nearest_value}
\title{Extract nearest value for single point}
\usage{
extract_nearest_value(x, point, max_range = NULL)
}
\arguments{
\item{x}{RasterLayer}

\item{point}{single sf object created by subsetting a single row from sf df.}

\item{max_range}{}
}
\value{
data.frame with extract value and distance from cell
}
\description{
For each NA, get cells in buffer, subset raster, calc distance, pick smallest
}
