% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{foxit}
\alias{foxit}
\title{Open PDF in Foxit with no lock}
\usage{
foxit(path)
}
\arguments{
\item{path}{path}
}
\value{
system
}
\description{
https://kb.foxitsoftware.com/hc/en-us/articles/360040661611-How-to-make-Foxit-PhantomPDF-Foxit-Reader-not-lock-PDF-file-when-open-it
}
