% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_metadata.R
\name{ggsaveme}
\alias{ggsaveme}
\title{GGsave metadata}
\usage{
ggsaveme(filename, plot = NULL, comment = NULL, ...)
}
\arguments{
\item{filename}{character Path to save file}

\item{plot}{Object to save. Last plot if unspecified}

\item{comment}{character}

\item{...}{additional arguments passed to ggsave}
}
\value{

}
\description{
Source is empty if the function was called from an unsaved script.
}
