% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nearest.R
\name{gp_ext_buff}
\alias{gp_ext_buff}
\title{Extract buffer, return 1 value}
\usage{
gp_ext_buff(points, rast, dist, coverage = 0.4, fct = median)
}
\arguments{
\item{points}{sf point object}

\item{rast}{raster* object}

\item{dist}{distance}

\item{coverage}{defaults to 0.4}

\item{fct}{median}
}
\value{
extracted and merged values
}
\description{
Extract buffer, return 1 value
}
