% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gp_ggaes}
\alias{gp_ggaes}
\title{gp_ggaes}
\usage{
gp_ggaes(geom)
}
\arguments{
\item{geom}{character which geom?}
}
\value{
vector of argument names
}
\description{
ggplot possible aesthetics
}
\details{
Lists all possible aesthetics arguments for a given geom
}
\examples{
gp_ggaes("violin")
}
