% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{gp_gplot}
\alias{gp_gplot}
\title{gplot from rasterVis hacked}
\usage{
gp_gplot(x, maxpixels = 50000, title = names(x)[1], filt_val = NULL)
}
\arguments{
\item{x}{a RasterLayer or SpatRaster}

\item{maxpixels}{Maximum number of pixels to use}

\item{filt_val}{filter value for categorical raster}
}
\value{
ggplot
}
\description{
see https://github.com/oscarperpinan/rastervis/blob/master/R/gplot.R
as opposed to the rasterVis implementation, this version does all the work
already of setting geom_raster, scale_fill and coord_fixed
}
\details{
Not implemented for RasterStack
}
