% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{gp_open_gmaps}
\alias{gp_open_gmaps}
\title{Open location in googlemaps}
\usage{
gp_open_gmaps(point)
}
\arguments{
\item{point}{sf point object}
}
\value{
open a new browser tab
}
\description{
Open location in googlemaps
}
\examples{
open_gmaps(point)
}
