% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{gp_point_neighbor}
\alias{gp_point_neighbor}
\title{Plot point neighborhood from raster}
\usage{
gp_point_neighbor(point, ras, dist = 8000)
}
\arguments{
\item{point}{only single sf point. will only consider first row}

\item{ras}{rasteLayer}

\item{dist}{range}
}
\value{
ggplot
}
\description{
Plot point neighborhood from raster
}
\examples{
gp_point_neibor(point, lcras_ori, 80000)
}
