% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gp_scale2}
\alias{gp_scale2}
\title{Scale2}
\usage{
gp_scale2(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector to scale}

\item{na.rm}{logical}
}
\value{

}
\description{
Same as scale but works with vectors, not matrices, and therefore with mutate
}
