% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{gp_stitch}
\alias{gp_stitch}
\title{A better knitr::stitch}
\usage{
gp_stitch(
  script = NULL,
  template = system.file("misc", "knitr-template.Rmd", package = "knitr"),
  output = NULL,
  text = NULL,
  envir = parent.frame(),
  rm_view = TRUE
)
}
\arguments{
\item{script}{Path to the R script. Default is the current document.}

\item{template}{Path of the template to use. By default, the Rnw template in
this package; there is also an HTML template in \pkg{knitr}.}

\item{output}{Output filename, passed to \code{\link{knit}}). By default,
the base filename of the script is used. @param template}

\item{text}{see knitr}

\item{envir}{see knitr}

\item{rm_view}{don't run View function calls}
}
\value{
output path
}
\description{
I had issues with stich which was behaving strangely and would not embed figures,
or would only work in the root folder (which is not what I wanted), or wouldn't
work with changed output file name. So here's a
modified version that runs knitr::stitch on the current script and saves the
html report to a stitch folder in the project.
}
\details{
NOTE there's something weird happening with the title, but whatever.
could use envir = new.env() in case there's an issue, but then might miss things from Rprofile
}
