% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gp_unwrap}
\alias{gp_unwrap}
\title{Unwrap string}
\usage{
gp_unwrap(x)
}
\arguments{
\item{x}{single character string}
}
\value{
character
}
\description{
Convenience function to allow writing text in source with line breaks,
saving character without.
}
\examples{
gp_unwrap("This is a very long string
          with spaces and line breaks.
          but that doesn't matter.")
}
