% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{gptools_include_path}
\alias{gptools_include_path}
\title{Get the gptools include path for compiling Stan programs with \code{cmdstanr} or \code{RStan}.}
\usage{
gptools_include_path()
}
\description{
Get the gptools include path for compiling Stan programs with \code{cmdstanr} or \code{RStan}.
}
