% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{hptmst}
\alias{hptmst}
\title{here-paste-tmst}
\usage{
hptmst(dir, main, ext, ...)
}
\arguments{
\item{dir}{folder}

\item{main}{character main name}

\item{ext}{character}

\item{...}{passed to tmst}
}
\value{
dunno
}
\description{
here-paste-tmst
}
