% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ignore}
\alias{ignore}
\title{ignore}
\usage{
ignore(...)
}
\arguments{
\item{...}{expression}
}
\description{
Can be used to mute view in script (e.g. for sourcing), or comment
out a block of code.
}
\examples{
view <- ignore
ignore({ expr })

}
