% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_whole}
\alias{is_whole}
\title{Check if integer}
\usage{
is_whole(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric}

\item{tol}{tolerance}
}
\value{
boolean
}
\description{
See is.wholenumber in http://web.mit.edu/~r/current/lib/R/library/base/html/integer.html
}
