% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{knit_w_tmst}
\alias{knit_w_tmst}
\title{Knit with timestamp}
\usage{
knit_w_tmst(input, ...)
}
\arguments{
\item{input}{input}

\item{...}{not used}
}
\value{
not sure
}
\description{
inspired by https://bookdown.org/yihui/rmarkdown-cookbook/custom-knit.html
This is to be used in the yaml of a Rmarkdown document with argument
knit: gptools::knit_w_tmst
}
