% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{last_tmst}
\alias{last_tmst}
\title{Read last timestamped RDS file}
\usage{
last_tmst(fold, pattern)
}
\arguments{
\item{pattern}{}
}
\value{
R object read from RDS file
}
\description{
Read last timestamped RDS file
}
