% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{list_cols_to_csv}
\alias{list_cols_to_csv}
\title{list_cols_to_csv}
\usage{
list_cols_to_csv(df, file_path, ...)
}
\arguments{
\item{df}{data.frame}

\item{file_path}{chr}

\item{...}{passed to write_excel_csv}
}
\description{
list_cols_to_csv
}
