% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{list_tmst}
\alias{list_tmst}
\title{List timestamped}
\usage{
list_tmst(fold, pattern = ".", recursive = FALSE)
}
\arguments{
\item{fold}{path}

\item{pattern}{pattern passed to list.files}
}
\value{
tibble
}
\description{
List timestamped files fitting a pattern with parsed date.
}
