% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahadist_map.R
\name{mahadist}
\alias{mahadist}
\title{Calculate mahalanobis distance}
\usage{
mahadist(dat, world, vars, xy = c("X", "Y"))
}
\arguments{
\item{xy}{}
}
\value{

}
\description{
Calculate mahalanobis distance
}
