% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahadist_map.R
\name{mahagrad}
\alias{mahagrad}
\title{Variant of mahalanobis map}
\usage{
mahagrad(dat, xy = c("X", "Y"), mask = F)
}
\arguments{
\item{dat}{data}

\item{xy}{coordinate columns}

\item{mask}{boolean}
}
\value{
ggplot
}
\description{
Uses gradient
}
