% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahadist_map.R
\name{make_grid}
\alias{make_grid}
\title{Make grid}
\usage{
make_grid(reso)
}
\arguments{
\item{reso}{in degree}
}
\value{
tibble
}
\description{
Specific use in spatial model predictions and mahalanobis analysis.
}
