% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{make_sf_wgs84}
\alias{make_sf_wgs84}
\title{Create sf object (WGS84) from dataframe}
\usage{
make_sf_wgs84(data, coords = NULL)
}
\arguments{
\item{data}{data}

\item{coords}{cloumn names like c("x", "y")}
}
\value{
sf object
}
\description{
Guesses which columns contain location information
}
