% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mk_hdr}
\alias{mk_hdr}
\title{Make header}
\usage{
mk_hdr(text, symbol = "*", width = 60, symbol2 = "-")
}
\arguments{
\item{text}{character}

\item{symbol}{single character}

\item{width}{numeric}
}
\description{
Creates a section header
Need to position the cursor after function call and run with alt+enter for best results.
}
\examples{
mk_hdr("My Title")
}
