% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{mk_timestamp}
\alias{mk_timestamp}
\title{mk_timestamp}
\usage{
mk_timestamp(file)
}
\arguments{
\item{file}{The file of interest}
}
\value{
file path with timestamp
}
\description{
adds a timestamp YYYYMMDD_HHMMSS to the file path
}
