% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{plot_corr}
\alias{plot_corr}
\title{Plot correlation}
\usage{
plot_corr(data, x, y)
}
\arguments{
\item{data}{data}

\item{x}{variable}

\item{y}{variable}
}
\value{
ggplot
}
\description{
Plot correlation
}
