% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_graphviz.R
\name{plot_lavaan}
\alias{plot_lavaan}
\title{Lavaan plot}
\usage{
plot_lavaan(model, layout = "dot", render = TRUE)
}
\arguments{
\item{model}{summary of the fit from lavaan}

\item{layout}{esthetics}

\item{render}{boolean}
}
\value{
?
}
\description{
Lavaan plot
}
