% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_graphviz.R
\name{plot_psem}
\alias{plot_psem}
\title{PSEM plot}
\usage{
plot_psem(model, layout = "dot", render = TRUE)
}
\arguments{
\item{model}{from psem}

\item{layout}{"dot" or "neato"}

\item{render}{boolean}
}
\value{
?
}
\description{
PSEM plot
}
