% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_graphviz.R
\name{psemtolav}
\alias{psemtolav}
\title{Helps to transform psem code to lavaan}
\usage{
psemtolav(string, data)
}
\arguments{
\item{string}{psem code}

\item{data}{data}
}
\value{
character
}
\description{
Helps to transform psem code to lavaan
}
