% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{pspls}
\alias{pspls}
\title{pull-sort-plot string input}
\usage{
pspls(data, col)
}
\arguments{
\item{data}{df}

\item{col}{name}
}
\value{
ggplot
}
\description{
might be easier to use with map
}
\examples{
pspl(iris, "Sepal.Length")
}
