% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{ptmst}
\alias{ptmst}
\title{paste-tmst}
\usage{
ptmst(main, ext, ...)
}
\arguments{
\item{main}{character main name}

\item{ext}{character}

\item{...}{passed to tmst}
}
\value{
path character
}
\description{
paste-tmst
}
