% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_metadata.R
\name{record_meta}
\alias{record_meta}
\title{Record metadata}
\usage{
record_meta(x, file, obj_name = NULL)
}
\arguments{
\item{x}{Object to save}

\item{file}{character Path to save file}

\item{obj_name}{only used if called internally}
}
\value{

}
\description{
Writes metadata to CSV file located in archd project folder. Tracks changes.
}
