% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_metadata.R
\name{saveme}
\alias{saveme}
\title{Save metadata}
\usage{
saveme(x, file, FUN = saveRDS, ...)
}
\arguments{
\item{x}{Object to save}

\item{file}{character Path to save file}

\item{FUN}{function used for saving. Default is saveRDS.}

\item{...}{additional arguments passed to FUN}
}
\value{

}
\description{
Wrapper function that saves an object using FUN and related metadata.
Supported functions include saveRDS(), write_csv() and write_excel_csv().
}
\details{
Doesn't currently work with ggsave.
}
