% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{show_heat}
\alias{show_heat}
\title{Make heat map}
\usage{
show_heat(df, rows)
}
\arguments{
\item{df}{data}

\item{rows}{character of column names}
}
\description{
Specify rows, columns will be all other numeric columns
}
