% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{tmst}
\alias{tmst}
\title{Timestamp}
\usage{
tmst(ext = NULL, time = T, prefix = "_c")
}
\arguments{
\item{ext}{character extension}

\item{time}{logical Should time (HMS) also be included?}

\item{prefix}{character to be added before, defaults to "_c"}
}
\value{
character squished timestamp
}
\description{
Squishes, especially useful for file names
}
