% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archd_fct.R
\name{update_log}
\alias{update_log}
\title{update_log}
\usage{
update_log(file, fold_path, method = NA)
}
\arguments{
\item{file}{file path}

\item{fold_path}{fold}

\item{method}{chr}
}
\description{
update_log
}
