% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wrnam}
\alias{wrnam}
\title{Write dataframe names}
\usage{
wrnam(x, sort = FALSE)
}
\arguments{
\item{x}{anything that has names}

\item{sort}{alphabetical sorting}
}
\value{
x
}
\description{
Convenience function
}
\examples{
wrnam(cars)
}
