% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_sim_df.R
\name{comp_for}
\alias{comp_for}
\title{comp_for}
\usage{
comp_for(df1, df2, id = NULL)
}
\arguments{
\item{id}{unique id column name maintained between dataframes}
}
\value{

}
\description{
Goes cell by cell, for all columns and rows
One issue is that the dfs need to have the same row (and column) order.
dplyr::arrange may not always solve the problem, and so sqldf provides a good alternative,
but doesn't give the row number.
}
