% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nearest.R
\name{extract_nearest_layer}
\alias{extract_nearest_layer}
\title{Extract nearest for RasterLayer}
\usage{
extract_nearest_layer(x, y, max_range = NULL)
}
\arguments{
\item{x}{RasterLayer}

\item{y}{points represented by a sf object}

\item{max_range}{numeric maximum distance in meters}
}
\value{
vector of values
}
\description{
Takes 1 layer, group same locations together
}
