% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{gp_pointmap}
\alias{gp_pointmap}
\title{Point map}
\usage{
gp_pointmap(
  data,
  id_col = NULL,
  coords = NULL,
  type = c("mapview", "leaflet", "ggplot")
)
}
\arguments{
\item{data}{dataset}

\item{coords}{numeric vector of length 2}

\item{type}{gg or leaf}

\item{ids}{character colum name to identify ids to plot}
}
\value{
ggplot or leaflet map
}
\description{
Can be used with df or sf, and implemented with ggplot, leaflet, and mapview
}
