% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{last_tmst}
\alias{last_tmst}
\title{Last timestamped}
\usage{
last_tmst(fold, pattern = "", load = TRUE, prev = 0)
}
\arguments{
\item{fold}{folder path}

\item{pattern}{regex pattern passed to list.files}

\item{load}{logical wanna load to file or just get it's name. use FALSE if the file is not rds format}

\item{prev}{int previous version before last}
}
\value{
R object read from RDS file
}
\description{
Read last timestamped (RDS) file
Default is to load file, but can return only name
}
