% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{last_tmst}
\alias{last_tmst}
\title{Read last timestamped RDS file}
\usage{
last_tmst(fold, pattern = "", load = TRUE)
}
\arguments{
\item{load}{logical wanna load to file or just get it's name. use FALSE if the file is not rds format}
}
\value{
R object read from RDS file
}
\description{
Read last timestamped RDS file
}
