% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahadist_map.R
\name{mahadist}
\alias{mahadist}
\title{Calculate mahalanobis distance}
\usage{
mahadist(dat, world, vars, xy = c("X", "Y"))
}
\arguments{
\item{dat}{data}

\item{world}{data.frame}

\item{vars}{variables used}

\item{xy}{coordinate column names as character vector}
}
\value{
input data.frame with added columns
}
\description{
Calculate mahalanobis distance
}
