% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{tmst}
\alias{tmst}
\title{Timestamp}
\usage{
tmst(ext = NULL, time = T, prefix = "_c")
}
\arguments{
\item{time}{logical Should time (HMS) also be included?}
}
\value{
character squishes timestamp
}
\description{
Squishes, especially useful for file names
}
